Moved SearchResultsTable out of Beta.
